requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
dj-database-url==2.0.0
Django==4.2.3
django-crispy-forms==2.0
gunicorn==21.2.0
mysql-connector==2.2.9
mysql-connector-python==8.1.0
packaging==23.1
Pillow==10.0.0
protobuf==4.21.12
psycopg2==2.9.6
pyodbc==4.0.39
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
